requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.2
cartopy==0.24.0
cffi==2.0.0
cftime==1.6.4
charset-normalizer==3.4.2
cryptography==46.0.3
docker==7.1.0
fastapi==0.122.0
filelock==3.13.1
Flask==3.1.2
flask-cors==6.0.1
frozenlist==1.7.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.2.4
gunicorn==23.0.0
h5py==3.16.0
h11==0.16.0
huey==2.5.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
llvmlite==0.44.0
Mako==1.3.10
matplotlib==3.10.3
mlflow==3.6.0
mlflow-skinny==3.6.0
mlflow-tracing==3.6.0
mpmath==1.3.0
multidict==6.6.3
munkres==1.1.4
netcdf4==1.7.2
networkx==3.3
numcodecs==0.16.1
numpy==2.1.2
opentelemetry-api==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
pandas==2.3.0
propcache==0.3.2
protobuf==6.33.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
PySide6==6.9.1
python-dotenv==1.2.1
pyyaml==6.0.2
requests==2.32.4
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.16.0
setuptools==80.9.0
SQLAlchemy==2.0.44
sqlparse==0.5.3
starlette==0.50.0
sympy==1.13.3
tqdm==4.67.1
triton==3.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
Werkzeug==3.1.3
xarray==2025.6.1
yarl==1.20.1
zarr==3.1.3
zstandard==0.23.0

# GPU PyTorch / related
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
torch-geometric==2.6.1

# VCS / editable installs
# -e git+https://github.com/SantanderMetGroup/deep4downscaling.git
